imageio: make imageinput_open_without_config() exception-safe
It is best to avoid exceptions so that the Rust layer can
translate errors into rust error types. Replace the exception throwing
with a test case that demonstrates how we will query for errors using
the sys crate.

The higher-level bindings can use these techniques to wrap the errors
in a Result type later. For now, we will eliminate the exceptions
to pave the way for those later changes.

Add a unit test that ensures that we can retrieve errors without
throwing exceptions.

Signed-off-by: David Aguilar <davvid@gmail.com>